+++ src/medicines/medicines.module.ts
@@ -1,4 +1,5 @@
 import { Module } from '@nestjs/common';
+import { MedicinesCacheService } from './cache/medicines-cache.service';
 import { MedicinesController } from './medicines.controller';
 import { MedicinesService } from './medicines.service';
 import { CnMedicinesService } from './sources/cn-medicines.service';
@@ -6,7 +7,12 @@ import { DrugbankMedicinesService } from './sources/drugbank-medicines.service';
 
 @Module({
   controllers: [MedicinesController],
-  providers: [MedicinesService, DrugbankMedicinesService, CnMedicinesService],
+  providers: [
+    MedicinesService,
+    MedicinesCacheService,
+    DrugbankMedicinesService,
+    CnMedicinesService,
+  ],
 })
 // eslint-disable-next-line @typescript-eslint/no-extraneous-class -- NestJS @Module requires a class declaration
 export class MedicinesModule {}